MiddleWare migrated to tf2. Added DnnModel class, this class instantiate the network components
